split off the cff-version specific bibtex_author behavior into its own files
